alx_travel_app/alx_travel_app/requirement.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.2
celery>=5.2
drf-yasg>=1.21
mysqlclient>=2.1
